+++ NEWS.md
@@ -1,3 +1,30 @@
+# ratios 1.2.0
+
+## Major changes
+
+Functions are partly renamed to leave the '.' mostly for generic functions
+
+* function ratio.ds is now called: ratioDT
+* function ratio.append_smalles is now called: ratio_append_smallest
+* function preparation.DT2 is now called: preparationDT2
+* Bug in correction for adhering particles of method 'M1' removed: Calculation of Errors is optional for functions 'Error = TRUE/FALSE'. 
+
+## Minor changes
+
+* at several occasions better testing for correct dimensions of input data sets
+* function select.VarsElements has new option 'invert = TRUE/FALSE'.
+* function ratioDT has new option 'id.vars'. If given, the function is faster.
+
+# ratios 1.1.1
+
+## Major changes
+
+* function ratio.DT1_DT2 is now called: ratio.ds (for "ratio of data sets").
+ratio.DT1_DT2 is kept as alias for backwards compatibility
+* the data set UpperCrust is used as default DT2_replace now only in the function CorrectionAdheringParticles, no longer in the function preparation.DT2
+* documentation for ratio.ds is improved
+
+
 # ratios 1.1.0
 
 ## Major changes